+++ jpa/boot-data-envers/src/main/java/com/example/envers/repositories/CustomerRepository.java
@@ -4,4 +4,5 @@ import com.example.envers.entities.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.history.RevisionRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Long>, RevisionRepository<Customer, Long, Long> {}
+public interface CustomerRepository
+        extends JpaRepository<Customer, Long>, RevisionRepository<Customer, Long, Integer> {}
